% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZILN_sup.R
\name{covar.transformation}
\alias{covar.transformation}
\title{Getting the transformed covariance matrix for one sample}
\usage{
covar.transformation(covar, n.tax)
}
\arguments{
\item{covar}{row covariance vector}

\item{n.tax}{Number of taxa. This is K + 1}
}
\description{
The covariance matrix for each sample is the kronecker product between the identity matrix
   with K dimensions (number of taxa - 1) and the covariance row vector (added 1 for intercept)
}
