% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key}
\alias{key}
\title{True covariate coefficients accompanying test data set \code{test.main}.}
\format{A data frame with 259 rows and 5 variables:
\describe{
  \item{Intercept}{Continuous}
  \item{covar1}{Continuous}
  \item{covar2}{Continuous}
  \item{covar3}{Continuous}
  \item{covar4}{Continuous}
}}
\usage{
key
}
\description{
A data set containing true coefficient values for all covariates (4 covariates) on
taxa (259 taxa). This dataset is simulated by Dr. Zhigang Li.
}
\keyword{datasets}
