% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZILN_main.R
\name{MZILN.cv}
\alias{MZILN.cv}
\title{Cross-validation for MZILN}
\usage{
MZILN.cv(log.data.full, matrix.data.full, n.folds, n.lam, lambda.values,
  reg.method = "mcp")
}
\description{
Does k-fold cross-validation for MZILN, produces a plot and returns a value for lambda
}
