% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZILN_sup.R
\name{MZILN.transformation}
\alias{MZILN.transformation}
\title{This function performs per sample taxonomic transformation}
\usage{
MZILN.transformation(df.main, df.covar, verbose = F)
}
\arguments{
\item{main}{A sample from the taxonomic abundance matrix}

\item{n.tax}{Number of total taxa}
}
\description{
For a sample, we find the log-ratio transformation of all non-zero taxa as well as the design matrix
}
