% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputcleaning.R
\name{rm.zeros}
\alias{rm.zeros}
\title{Removing zero rows and columns}
\usage{
rm.zeros(data)
}
\arguments{
\item{data}{The input relative abundance data frame}
}
\value{
return the cleaned data
}
\description{
Removing zero rows and columns
}
\examples{
\dontrun{
rm.zeros(df.result)
}
}
