% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.covariates}
\alias{test.covariates}
\title{Simulated data frame of patient covariates accompanying \code{test.main}.}
\format{A data frame with 200 rows and 4 variables:
\describe{
  \item{covar1}{Continuous}
  \item{covar2}{Continuous}
  \item{covar3}{Continuous}
  \item{covar4}{Discrete}
}}
\usage{
test.covariates
}
\description{
A dataset containing covariate data (4 covariates) for
200 patients. This dataset is simulated by Dr. Zhigang Li.
}
\keyword{datasets}
