% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.main}
\alias{test.main}
\title{Microbiome Relative Abundance data for 200 patients}
\format{A data frame with 200 rows and 260 columns:
\describe{
   \item{w1} Taxa w1
   \item{w2} Taxa w2
}}
\usage{
test.main
}
\description{
Test data file containing simulated relative abundance data from ZILN paper
   prepared by Dr. Zhigang Li with 200 patients.
}
\keyword{datasets}
