% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZILN.R
\name{ZILN.main}
\alias{ZILN.main}
\title{ZILN (Zero inflated Logistic Normal Regression)}
\usage{
ZILN.main(df.main, df.covar, covariates = colnames(df.covar), n.lam = 200,
  lam.min.ratio = 1e-04, reg.method = "mcp", n.folds = 5)
}
\arguments{
\item{df.main}{Data frame of patient relative abundance data}

\item{df.covar}{Data frame of patient covariates}

\item{covariates}{Specific covariates of interest. The default is all covariates
in \code{df.covar}}

\item{n.lam}{Number of \code{lambda} values to control the regularization as documented
in \code{PICASSO}. The default is 200.}

\item{lam.min.ratio}{Ratio between the upper and lower bound of generated \code{lambda} values.
The default is 0.0001.}

\item{reg.method}{The method of regularization. The default is \code{mcp}. Other options
include \code{lasso} and \code{scad}}

\item{n.folds}{The number of cross-validation folds. The default is 5.}
}
\value{
A list containing: \code{results} - a data frame of regression coefficients for each
    covariate of interest (as columns) per taxa (as rows); and \code{crossvalidation} - a data
    frame of each \code{lambda} value and the mean MSE of each value across cross validation folds
}
\description{
ZILN is based upon the Zero Inflated Logistic Normal Regression
for microbiome relative abundance data prepared by Dr. Zhigang Li.
}
