% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{features.confirm}
\alias{features.confirm}
\title{Confirmation of data features}
\usage{
features.confirm(df.main, df.covar, covariates = covar.name(df.covar))
}
\arguments{
\item{df.main}{Data frame containing Relative Abundances for each sample.}

\item{df.covar}{Data frame containing covariance data for each sample.}

\item{covariates}{Names of covariates of interest. The default is all covariates
in \code{df.covar}}
}
\value{
Relevant information about the input printed to the console which include: data format,
    concordance between data frames, number of subjects, 10 subject IDs, total input
    covariates, covariates of interest, Number of taxa, first 10 taxa names, number of zero subjects,
    number of zero taxa.
}
\description{
Confirmation of data features
}
\details{
MZILN generally takes three main arguments: First, a relative abundance
    of all taxa per subject, which is \code{df.main}. Second, a separate
    data frame containing all covariates by subject, which is \code{df.covar}.
    Third, an optional vector containing the names of a subset of covariates that is
    of specific interest \code{covariates}. The default mode for \code{covariates}
    is the entire columns of covariates in the data frame. ZILN expects subject ids to
    be the rownames of the data frame and each column should represent either a taxa's
    RA or a covariate value. This function checks all of those required features, display
    them as well as check whether or not the formatting is correct.
}
\examples{
#loading data frames
data(test.main)
data(test.covariates)
features.confirm(test.main, test.covariates, covariates = c("covar1","covar4"))
}
