% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.covariates}
\alias{test.covariates}
\title{Covariate data for 200 patients}
\format{A data frame with 200 rows and 4 columns
\describe{
   \item{covar1} Covariate covar1
   \item{covar2} Covariate covar2
}}
\usage{
test.covariates
}
\description{
Test file containing simulated data of 4 hypothetical covariates from ZILN paper
    prepared by Dr. Zhigang Li with 200 patients.
}
\keyword{datasets}
