% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.main}
\alias{test.main}
\title{Simulated data frame of microbiome relative abundance.}
\format{A data frame with 200 rows and 260 variables:
\describe{
  \item{taxa1}{Relative abundance}
  \item{taxa2}{Relative abundance}
  ...
}}
\usage{
test.main
}
\description{
A dataset containing the relative abundance of 260 taxa (as columns) for
200 patients. This dataset is simulated by Dr. Zhigang Li.
}
\keyword{datasets}
