% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.main}
\alias{test.main}
\title{Simulated data frame of microbiome relative abundance.}
\format{A data frame with 200 rows and 260 variables:
\describe{
  \item{taxa1}{Relative Abundance}
  \item{taxa2}{Relative Abundance}
  \item{taxa3}{Relative Abundance}
  \item{taxa4}{Relative Abundance}
  \item{taxa5}{Relative Abundance}
  \item{taxa6}{Relative Abundance}
  \item{taxa7}{Relative Abundance}
  \item{taxa8}{Relative Abundance}
  \item{taxa9}{Relative Abundance}
  \item{taxa10}{Relative Abundance}
  \item{taxa11}{Relative Abundance}
  \item{taxa12}{Relative Abundance}
  \item{taxa13}{Relative Abundance}
  \item{taxa14}{Relative Abundance}
  \item{taxa15}{Relative Abundance}
  \item{taxa16}{Relative Abundance}
  \item{taxa17}{Relative Abundance}
  \item{taxa18}{Relative Abundance}
  \item{taxa19}{Relative Abundance}
  \item{taxa20}{Relative Abundance}
  \item{taxa21}{Relative Abundance}
  \item{taxa22}{Relative Abundance}
  \item{taxa23}{Relative Abundance}
  \item{taxa24}{Relative Abundance}
  \item{taxa25}{Relative Abundance}
  \item{taxa26}{Relative Abundance}
  \item{taxa27}{Relative Abundance}
  \item{taxa28}{Relative Abundance}
  \item{taxa29}{Relative Abundance}
  \item{taxa30}{Relative Abundance}
  \item{taxa31}{Relative Abundance}
  \item{taxa32}{Relative Abundance}
  \item{taxa33}{Relative Abundance}
  \item{taxa34}{Relative Abundance}
  \item{taxa35}{Relative Abundance}
  \item{taxa36}{Relative Abundance}
  \item{taxa37}{Relative Abundance}
  \item{taxa38}{Relative Abundance}
  \item{taxa39}{Relative Abundance}
  \item{taxa40}{Relative Abundance}
  \item{taxa41}{Relative Abundance}
  \item{taxa42}{Relative Abundance}
  \item{taxa43}{Relative Abundance}
  \item{taxa44}{Relative Abundance}
  \item{taxa45}{Relative Abundance}
  \item{taxa46}{Relative Abundance}
  \item{taxa47}{Relative Abundance}
  \item{taxa48}{Relative Abundance}
  \item{taxa49}{Relative Abundance}
  \item{taxa50}{Relative Abundance}
  \item{taxa51}{Relative Abundance}
  \item{taxa52}{Relative Abundance}
  \item{taxa53}{Relative Abundance}
  \item{taxa54}{Relative Abundance}
  \item{taxa55}{Relative Abundance}
  \item{taxa56}{Relative Abundance}
  \item{taxa57}{Relative Abundance}
  \item{taxa58}{Relative Abundance}
  \item{taxa59}{Relative Abundance}
  \item{taxa60}{Relative Abundance}
  \item{taxa61}{Relative Abundance}
  \item{taxa62}{Relative Abundance}
  \item{taxa63}{Relative Abundance}
  \item{taxa64}{Relative Abundance}
  \item{taxa65}{Relative Abundance}
  \item{taxa66}{Relative Abundance}
  \item{taxa67}{Relative Abundance}
  \item{taxa68}{Relative Abundance}
  \item{taxa69}{Relative Abundance}
  \item{taxa70}{Relative Abundance}
  \item{taxa71}{Relative Abundance}
  \item{taxa72}{Relative Abundance}
  \item{taxa73}{Relative Abundance}
  \item{taxa74}{Relative Abundance}
  \item{taxa75}{Relative Abundance}
  \item{taxa76}{Relative Abundance}
  \item{taxa77}{Relative Abundance}
  \item{taxa78}{Relative Abundance}
  \item{taxa79}{Relative Abundance}
  \item{taxa80}{Relative Abundance}
  \item{taxa81}{Relative Abundance}
  \item{taxa82}{Relative Abundance}
  \item{taxa83}{Relative Abundance}
  \item{taxa84}{Relative Abundance}
  \item{taxa85}{Relative Abundance}
  \item{taxa86}{Relative Abundance}
  \item{taxa87}{Relative Abundance}
  \item{taxa88}{Relative Abundance}
  \item{taxa89}{Relative Abundance}
  \item{taxa90}{Relative Abundance}
  \item{taxa91}{Relative Abundance}
  \item{taxa92}{Relative Abundance}
  \item{taxa93}{Relative Abundance}
  \item{taxa94}{Relative Abundance}
  \item{taxa95}{Relative Abundance}
  \item{taxa96}{Relative Abundance}
  \item{taxa97}{Relative Abundance}
  \item{taxa98}{Relative Abundance}
  \item{taxa99}{Relative Abundance}
  \item{taxa100}{Relative Abundance}
  \item{taxa101}{Relative Abundance}
  \item{taxa102}{Relative Abundance}
  \item{taxa103}{Relative Abundance}
  \item{taxa104}{Relative Abundance}
  \item{taxa105}{Relative Abundance}
  \item{taxa106}{Relative Abundance}
  \item{taxa107}{Relative Abundance}
  \item{taxa108}{Relative Abundance}
  \item{taxa109}{Relative Abundance}
  \item{taxa110}{Relative Abundance}
  \item{taxa111}{Relative Abundance}
  \item{taxa112}{Relative Abundance}
  \item{taxa113}{Relative Abundance}
  \item{taxa114}{Relative Abundance}
  \item{taxa115}{Relative Abundance}
  \item{taxa116}{Relative Abundance}
  \item{taxa117}{Relative Abundance}
  \item{taxa118}{Relative Abundance}
  \item{taxa119}{Relative Abundance}
  \item{taxa120}{Relative Abundance}
  \item{taxa121}{Relative Abundance}
  \item{taxa122}{Relative Abundance}
  \item{taxa123}{Relative Abundance}
  \item{taxa124}{Relative Abundance}
  \item{taxa125}{Relative Abundance}
  \item{taxa126}{Relative Abundance}
  \item{taxa127}{Relative Abundance}
  \item{taxa128}{Relative Abundance}
  \item{taxa129}{Relative Abundance}
  \item{taxa130}{Relative Abundance}
  \item{taxa131}{Relative Abundance}
  \item{taxa132}{Relative Abundance}
  \item{taxa133}{Relative Abundance}
  \item{taxa134}{Relative Abundance}
  \item{taxa135}{Relative Abundance}
  \item{taxa136}{Relative Abundance}
  \item{taxa137}{Relative Abundance}
  \item{taxa138}{Relative Abundance}
  \item{taxa139}{Relative Abundance}
  \item{taxa140}{Relative Abundance}
  \item{taxa141}{Relative Abundance}
  \item{taxa142}{Relative Abundance}
  \item{taxa143}{Relative Abundance}
  \item{taxa144}{Relative Abundance}
  \item{taxa145}{Relative Abundance}
  \item{taxa146}{Relative Abundance}
  \item{taxa147}{Relative Abundance}
  \item{taxa148}{Relative Abundance}
  \item{taxa149}{Relative Abundance}
  \item{taxa150}{Relative Abundance}
  \item{taxa151}{Relative Abundance}
  \item{taxa152}{Relative Abundance}
  \item{taxa153}{Relative Abundance}
  \item{taxa154}{Relative Abundance}
  \item{taxa155}{Relative Abundance}
  \item{taxa156}{Relative Abundance}
  \item{taxa157}{Relative Abundance}
  \item{taxa158}{Relative Abundance}
  \item{taxa159}{Relative Abundance}
  \item{taxa160}{Relative Abundance}
  \item{taxa161}{Relative Abundance}
  \item{taxa162}{Relative Abundance}
  \item{taxa163}{Relative Abundance}
  \item{taxa164}{Relative Abundance}
  \item{taxa165}{Relative Abundance}
  \item{taxa166}{Relative Abundance}
  \item{taxa167}{Relative Abundance}
  \item{taxa168}{Relative Abundance}
  \item{taxa169}{Relative Abundance}
  \item{taxa170}{Relative Abundance}
  \item{taxa171}{Relative Abundance}
  \item{taxa172}{Relative Abundance}
  \item{taxa173}{Relative Abundance}
  \item{taxa174}{Relative Abundance}
  \item{taxa175}{Relative Abundance}
  \item{taxa176}{Relative Abundance}
  \item{taxa177}{Relative Abundance}
  \item{taxa178}{Relative Abundance}
  \item{taxa179}{Relative Abundance}
  \item{taxa180}{Relative Abundance}
  \item{taxa181}{Relative Abundance}
  \item{taxa182}{Relative Abundance}
  \item{taxa183}{Relative Abundance}
  \item{taxa184}{Relative Abundance}
  \item{taxa185}{Relative Abundance}
  \item{taxa186}{Relative Abundance}
  \item{taxa187}{Relative Abundance}
  \item{taxa188}{Relative Abundance}
  \item{taxa189}{Relative Abundance}
  \item{taxa190}{Relative Abundance}
  \item{taxa191}{Relative Abundance}
  \item{taxa192}{Relative Abundance}
  \item{taxa193}{Relative Abundance}
  \item{taxa194}{Relative Abundance}
  \item{taxa195}{Relative Abundance}
  \item{taxa196}{Relative Abundance}
  \item{taxa197}{Relative Abundance}
  \item{taxa198}{Relative Abundance}
  \item{taxa199}{Relative Abundance}
  \item{taxa200}{Relative Abundance}
  \item{taxa201}{Relative Abundance}
  \item{taxa202}{Relative Abundance}
  \item{taxa203}{Relative Abundance}
  \item{taxa204}{Relative Abundance}
  \item{taxa205}{Relative Abundance}
  \item{taxa206}{Relative Abundance}
  \item{taxa207}{Relative Abundance}
  \item{taxa208}{Relative Abundance}
  \item{taxa209}{Relative Abundance}
  \item{taxa210}{Relative Abundance}
  \item{taxa211}{Relative Abundance}
  \item{taxa212}{Relative Abundance}
  \item{taxa213}{Relative Abundance}
  \item{taxa214}{Relative Abundance}
  \item{taxa215}{Relative Abundance}
  \item{taxa216}{Relative Abundance}
  \item{taxa217}{Relative Abundance}
  \item{taxa218}{Relative Abundance}
  \item{taxa219}{Relative Abundance}
  \item{taxa220}{Relative Abundance}
  \item{taxa221}{Relative Abundance}
  \item{taxa222}{Relative Abundance}
  \item{taxa223}{Relative Abundance}
  \item{taxa224}{Relative Abundance}
  \item{taxa225}{Relative Abundance}
  \item{taxa226}{Relative Abundance}
  \item{taxa227}{Relative Abundance}
  \item{taxa228}{Relative Abundance}
  \item{taxa229}{Relative Abundance}
  \item{taxa230}{Relative Abundance}
  \item{taxa231}{Relative Abundance}
  \item{taxa232}{Relative Abundance}
  \item{taxa233}{Relative Abundance}
  \item{taxa234}{Relative Abundance}
  \item{taxa235}{Relative Abundance}
  \item{taxa236}{Relative Abundance}
  \item{taxa237}{Relative Abundance}
  \item{taxa238}{Relative Abundance}
  \item{taxa239}{Relative Abundance}
  \item{taxa240}{Relative Abundance}
  \item{taxa241}{Relative Abundance}
  \item{taxa242}{Relative Abundance}
  \item{taxa243}{Relative Abundance}
  \item{taxa244}{Relative Abundance}
  \item{taxa245}{Relative Abundance}
  \item{taxa246}{Relative Abundance}
  \item{taxa247}{Relative Abundance}
  \item{taxa248}{Relative Abundance}
  \item{taxa249}{Relative Abundance}
  \item{taxa250}{Relative Abundance}
  \item{taxa251}{Relative Abundance}
  \item{taxa252}{Relative Abundance}
  \item{taxa253}{Relative Abundance}
  \item{taxa254}{Relative Abundance}
  \item{taxa255}{Relative Abundance}
  \item{taxa256}{Relative Abundance}
  \item{taxa257}{Relative Abundance}
  \item{taxa258}{Relative Abundance}
  \item{taxa259}{Relative Abundance}
  \item{taxa260}{Relative Abundance}
}}
\usage{
test.main
}
\description{
A dataset containing the relative abundance of 260 taxa (as columns) for
200 patients. This dataset is simulated by Dr. Zhigang Li.
}
\keyword{datasets}
